Create a custom transport network used for routing in R5 with modified OSM car speeds

The function builds a transport network with modified OSM car speeds. The function has two modes, hereby refereed to as "edge mode" and "polygon mode". See the Details section below.

Usage

build_custom_network(
  data_path,
  new_carspeeds,
  output_path = tempdir_unique(),
  default_speed = NULL,
  percentage_mode = TRUE,
  verbose = FALSE,
  elevation = "TOBLER"
)

Arguments

data_path

Character. Path to the directory with the.pbf file of the OSM network and optionally supporting data such as elevation and GTFS.

new_carspeeds

A data.frame specifying the speed modifications. The table must contain columns osm_id and max_speed. OR A sf data.frame specifying the speed modifications. The table must contain columns sf polygon, scale, priority. See details.

output_path

Character. A string pointing to the directory where the modified network.dat will be saved. Must be different from data_path. Defaults to a temporary directory.

default_speed

Numeric. Default speed to use for road segments not specified in new_carspeeds. Must be >= 0. Defaults to NULL so that roads not listed have their speeds unchanged. When set to 0, the road segment is assumed to be closed.

percentage_mode

Logical. Only OSM mode. If TRUE, values in max_speed are interpreted as percentages of original speeds; if FALSE, as absolute speeds (km/h). Defaults to TRUE - percentages.

verbose

A logical. Whether to show R5 informative messages when running the function. Defaults to FALSE (please note that in such case R5 error messages are still shown). Setting verbose to TRUE shows detailed output, which can be useful for debugging issues not caught by r5r.

elevation

A string. The name of the impedance function to be used to calculate impedance for walking and cycling based on street slopes. Available options include TOBLER (Default) and MINETTI, or NONE to ignore elevation. R5 loads elevation data from .tif files saved inside the data_path directory. See more info in the Details section below.

Value

A r5r_network object representing the built network to connect with R5 routing engine, and a network.dat file saved in the output_path.

Details

Edge mode: The new speed factors must be passed as a data.frame indicating the new max speed of each OSM edge id. The table must contain columns osm_id and max_speed. Values in max_speed can be interpreted as percentages of original speeds or as absolute speeds (km/h) by using percentage_mode parameter.

Polygon mode: The new speed factors must be passed through "congestion region(s)" specified as an sf data.frame containing in each row an sf polygon outlining the "congestion region", scale and priority. Values in scale can only be interpreted as percentages of original speeds as percentage_mode must be TRUE. Priority is used to apply a hierarchy in case of overlapping polygons.

Elevation

More information about the TOBLER and MINETTI options to calculate the effects of elevation on travel times can be found in the references below:

Campbell, M. J., et al (2019). Using crowdsourced fitness tracker data to model the relationship between slope and travel rates. Applied geography, 106, 93-107. doi:10.1016/j.apgeog.2019.03.008 .

Minetti, A. E., et al (2002). Energy cost of walking and running at extreme uphill and downhill slopes. Journal of applied physiology. doi:10.1152/japplphysiol.01177.2001 .

Tobler, W. (1993). Three presentations on geographical analysis and modeling: Non-isotropic geographic modeling speculations on the geometry of geography global spatial analysis. Technical Report. National center for geographic information and analysis. 93 (1). https://escholarship.org/uc/item/05r820mz.

See also

Examples

library(r5r)

# path to OSM .pbf file
data_path <- system.file("extdata/poa", package = "r5r")

# data.frame with new speed info
edge_speeds <- read.csv(file.path(data_path, "poa_osm_congestion.csv"))

r5r_network_new_speeds <- r5r::build_custom_network(
  data_path = data_path,
  new_carspeeds = edge_speeds,
  output_path = tempdir(),
  percentage_mode = TRUE
)
#> ℹ Copying 3 files:
#>   /home/runner/work/_temp/Library/r5r/extdata/poa/poa_elevation.tif,
#>   /home/runner/work/_temp/Library/r5r/extdata/poa/poa_eptc.zip, and
#>   /home/runner/work/_temp/Library/r5r/extdata/poa/poa_trensurb.zip to
#>   /tmp/RtmpbTqFCr
#> ℹ Building network modifier...
#> ℹ Building new network...
#> ✔ Finished building network at /tmp/RtmpbTqFCr

# sf with congestion polygons
congestion_poly <- readRDS(file.path(data_path, "poa_poly_congestion.rds"))

r5r_network_congestion <- r5r::build_custom_network(
  data_path = data_path,
  new_carspeeds = congestion_poly,
  output_path = tempdir(),
  percentage_mode = TRUE
)
#> ℹ Copying 3 files:
#>   /home/runner/work/_temp/Library/r5r/extdata/poa/poa_elevation.tif,
#>   /home/runner/work/_temp/Library/r5r/extdata/poa/poa_eptc.zip, and
#>   /home/runner/work/_temp/Library/r5r/extdata/poa/poa_trensurb.zip to
#>   /tmp/RtmpbTqFCr
#> ℹ Deleting existing /tmp/RtmpbTqFCr/congested-poa_osm.pbf from the output
#>   folder.
#> ✔ Finished building network at /tmp/RtmpbTqFCr