Create a transport network used for routing in R5 with modified OSM car speeds

The function builds a transport network with modified OSM car speeds. The new speed factors must me passed from a .csv file indication the new max speed of each OSM edge id.

Usage

modify_osm_carspeeds(
  pbf_path,
  csv_path,
  output_dir = tempdir_unique(),
  default_speed = NULL,
  percentage_mode = TRUE,
  verbose = FALSE
)

Arguments

pbf_path

Character. Path to the .pbf file of the OSM network.

csv_path

Character. Path to the CSV file witha a table specifying the speed modifications. The table must contain columns osm_id and max_speed.

output_dir

Character. Directory where the modified network will be written. Defaults to a temporary directory.

default_speed

Numeric. Default speed to use for segments not specified in the CSV. Must be >= 0. Defaults to NULL so that roads not listed in. the CSV have their speeds unchanged.

percentage_mode

Logical. If TRUE, values in max_speed are interpreted as percentages of original speeds; if FALSE, as absolute speeds (km/h). Defaults to TRUE - percentages.

verbose

Logical. If TRUE, the function modifies the network verbosely and creates a verbose R5RCore. Defaults to FALSE.

Value

An R5 core object representing the rebuilt network with modified car speeds.

Details

The CSV must have columns named osm_id and max_speed. max_speed can be specified as a percentage of the original road speed or as an absolute speed in km/h. The function rebuilds the network in output_dir and returns a new r5r_network object.

Examples

library(r5r)

# path to OSM .pbf file
pbf_path <- system.file("extdata/poa/poa_osm.pbf", package = "r5r")

# path to CSV with a table pointing to the new speed info
speeds_csv_path <- system.file("extdata/poa/poa_osm_congestion.csv", package = "r5r")

r5r_network_new_speed <- r5r::modify_osm_carspeeds(
  pbf_path = pbf_path,
  csv_path = speeds_csv_path,
  output_dir = tempdir(),
  percentage_mode = TRUE
)
#> No raster .tif files found. Using elevation = 'NONE'.
#> 
#> No public transport data (gtfs) provided. Graph will be built with the street network only.
#> 
#> Finished building network.dat at /tmp/RtmpozMGr2/network.dat
#> New car network with modified speeds built at /tmp/RtmpozMGr2