r5r: Rapid Realistic Routing with 'R5'

Rapid realistic routing on multimodal transport networks (walk, bike, public transport and car) using R5, the Rapid Realistic Routing on Real-world and Reimagined networks engine https://github.com/conveyal/r5. The package allows users to generate detailed routing analysis or calculate travel time matrices using seamless parallel computing on top of the R5 Java machine. While R5 is developed by Conveyal, the package r5r is independently developed by a team at the Institute for Applied Economic Research (Ipea) with contributions from collaborators. Apart from the documentation in this package, users will find additional information on R5 documentation at https://docs.conveyal.com/. Although we try to keep new releases of r5r in synchrony with R5, the development of R5 follows Conveyal's independent update process. Hence, users should confirm if the R5 version implied by the Conveyal user manual (see https://docs.conveyal.com/changelog) corresponds with the R5 version that r5r depends on.

Usage

Please check the vignettes on the website.

See also

Author

Maintainer: Rafael H. M. Pereira rafa.pereira.br@gmail.com (ORCID)

Authors:

Marcus Saraiva marcus.saraiva@gmail.com (ORCID)

Daniel Herszenhut dhersz@gmail.com (ORCID)

Carlos Kaue Vieira Braga kaue@kauebraga.dev (ORCID)

Matthew Wigginton Conway mwconway@asu.edu (ORCID)

Other contributors:

Ipea - Institute for Applied Economic Research [copyright holder, funder]