Set the fare structure used when calculating transit fares

Sets the fare structure used by our "generic" fare calculator. A value of NULL is passed to fare_structure by the upstream routing and accessibility functions when fares are not to be calculated.

Usage

set_fare_structure(r5r_core, fare_structure)

Arguments

r5r_core

An object to connect with the R5 routing engine, created with setup_r5().

fare_structure

A fare structure object, following the convention set in setup_fare_structure(). This object describes how transit fares should be calculated. Please see the fare structure vignette to understand how this object is structured: vignette("fare_structure", package = "r5r").

Value

Invisibly returns TRUE.

See also