Set max Level of Transit Stress (LTS)

Set max Level of Transit Stress (LTS)

set_max_lts(r5r_core, max_lts)

Arguments

r5r_core

An object to connect with the R5 routing engine, created with setup_r5().

max_lts

A number (between 1 and 4). The maximum level of traffic stress that cyclists will tolerate. A value of 1 means cyclists will only travel through the quietest streets, while a value of 4 indicates cyclists can travel through any road.

Value

Invisibly returns TRUE.

See also