Set suboptimal minutes

Sets the number of suboptimal minutes considered in detailed_itineraries() routing. From R5 documentation: "This parameter compensates for the fact that GTFS does not contain information about schedule deviation (lateness). The min-max travel time range for some trains is zero, since the trips are reported to always have the same timings in the schedule. Such an option does not overlap (temporally) its alternatives, and is too easily eliminated by an alternative that is only marginally better. We want to effectively push the max travel time of alternatives out a bit to account for the fact that they don't always run on schedule".

Usage

set_suboptimal_minutes(
  r5r_core,
  suboptimal_minutes,
  fare_structure,
  shortest_path
)

Arguments

r5r_core

An object to connect with the R5 routing engine, created with setup_r5().

suboptimal_minutes

A number.

fare_structure

A fare structure object, following the convention set in setup_fare_structure(). This object describes how transit fares should be calculated. Please see the fare structure vignette to understand how this object is structured: vignette("fare_structure", package = "r5r").

shortest_path

A logical.

Value

Invisibly returns TRUE.

See also