Download R5 Jar file

Download a compiled JAR file of R5 and saves it locally. The JAR file is saved within the package directory. The package uses a compilation of R5 tailored for the purposes of r5r that keeps R5's essential features. Source code available at https://github.com/ipeaGIT/r5r.

download_r5(
  version = "6.0.1",
  quiet = FALSE,
  force_update = FALSE,
  temp_dir = FALSE
)

Arguments

version
character string with the version of R5 to be downloaded. Defaults to latest version '1.0'.

quiet
logical, passed to download.file. Defaults to FALSE

force_update
logical, Replaces the jar file stored locally with a new one. Defaults to FALSE.

temp_dir
logical, whether the R5 Jar file should be saved in temporary directory. Defaults to FALSE

Value

A jar file is saved locally in the r5r package directory

See also

Examples

if (FALSE) {

library(r5r)

download_r5(version = "6.0.1", temp_dir = TRUE)
}