Generate an r5r situation report to help debug errors

The function reports a list with the following information:

The package version of {r5r} in use.

The installed version of R5.jar.

The Java version in use.

The amount of memory set to Java through the java.parameters option.

The user's Session Info.

Usage

r5r_sitrep()

Value

A list with information of the versions of the r5r package, Java and R5 Jar in use, the memory set to Java and user's Session Info.

Examples

r5r_sitrep()
#> $r5r_package_version
#> [1] ‘2.0’
#> 
#> $r5_jar_version
#> [1] "7.1"
#> 
#> $java_version
#> [1] "21.0.2"
#> 
#> $set_memory
#> [1] "-Xmx2G"
#> 
#> $session_info
#> R version 4.4.2 (2024-10-31 ucrt)
#> Platform: x86_64-w64-mingw32/x64
#> Running under: Windows Server 2022 x64 (build 20348)
#> 
#> Matrix products: default
#> 
#> 
#> locale:
#> [1] LC_COLLATE=C                       LC_CTYPE=Portuguese_Brazil.utf8   
#> [3] LC_MONETARY=Portuguese_Brazil.utf8 LC_NUMERIC=C                      
#> [5] LC_TIME=Portuguese_Brazil.utf8    
#> 
#> time zone: America/Sao_Paulo
#> tzcode source: internal
#> 
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base     
#> 
#> other attached packages:
#> [1] r5r_2.0
#> 
#> loaded via a namespace (and not attached):
#>   [1] Rdpack_2.6.2             DBI_1.2.3                deldir_2.0-4            
#>   [4] httr2_1.0.7              s2_1.1.7                 remotes_2.5.0           
#>   [7] testthat_3.2.3           rlang_1.1.5              magrittr_2.0.3          
#>  [10] e1071_1.7-16             compiler_4.4.2           sfheaders_0.4.4         
#>  [13] systemfonts_1.1.0        callr_3.7.6              vctrs_0.6.5             
#>  [16] profvis_0.4.0            pkgconfig_2.0.3          h3jsr_1.3.1             
#>  [19] wk_0.9.4                 fastmap_1.2.0            backports_1.5.0         
#>  [22] ellipsis_0.3.2           fontawesome_0.5.3        geojsonsf_2.0.3         
#>  [25] promises_1.3.2           rmarkdown_2.29           sessioninfo_1.2.2       
#>  [28] ps_1.8.1                 ragg_1.3.3               purrr_1.0.2             
#>  [31] xfun_0.50                cachem_1.1.0             jsonlite_1.8.9          
#>  [34] pak_0.8.0                later_1.4.1              R6_2.5.1                
#>  [37] bslib_0.8.0              pkgload_1.4.0            brio_1.1.5              
#>  [40] jquerylib_0.1.4          Rcpp_1.0.14              knitr_1.49              
#>  [43] usethis_3.1.0            httpuv_1.6.15            tidyselect_1.2.1        
#>  [46] rstudioapi_0.17.1        yaml_2.3.10              miniUI_0.1.1.1          
#>  [49] curl_6.1.0               processx_3.8.5           pkgbuild_1.4.6          
#>  [52] tibble_3.2.1             shiny_1.9.1              withr_3.0.2             
#>  [55] evaluate_1.0.3           accessibility_1.4.0.9000 desc_1.4.3              
#>  [58] sf_1.0-19                rJava_1.0-11             units_0.8-5             
#>  [61] proxy_0.4-27             urlchecker_1.0.1         zip_2.3.1               
#>  [64] xml2_1.3.6               pillar_1.10.1            whisker_0.4.1           
#>  [67] KernSmooth_2.23-26       checkmate_2.3.2          generics_0.1.3          
#>  [70] ggplot2_3.5.1            munsell_0.5.1            scales_1.3.0            
#>  [73] xtable_1.8-4             gtfstools_1.4.0          class_7.3-23            
#>  [76] glue_1.8.0               tools_4.4.2              interp_1.1-6            
#>  [79] data.table_1.16.4        fs_1.6.5                 grid_4.4.2              
#>  [82] tidyr_1.3.1              rbibutils_2.3            devtools_2.4.5          
#>  [85] colorspace_2.1-1         patchwork_1.3.0          geobr_1.9.1             
#>  [88] cli_3.6.3                rappdirs_0.3.3           textshaping_0.4.0       
#>  [91] fansi_1.0.6              dplyr_1.1.4              concaveman_1.1.0        
#>  [94] downlit_0.4.4            V8_6.0.0                 gtable_0.3.6            
#>  [97] sass_0.4.9               digest_0.6.37            classInt_0.4-11         
#> [100] htmlwidgets_1.6.4        farver_2.1.2             memoise_2.0.1           
#> [103] htmltools_0.5.8.1        pkgdown_2.1.1.9000       lifecycle_1.0.4         
#> [106] mime_0.12               
#>