Generate an r5r situation report to help debug errors

The function reports a list with the following information:

The package version of {r5r} in use.

The installed version of R5.jar.

The Java version in use.

The amount of memory set to Java through the java.parameters option.

The user's Session Info.

r5r_sitrep()

Value

A list with information of the versions of the r5r package, Java and R5 Jar in use, the memory set to Java and user's Session Info.

Examples

r5r_sitrep()
#> $r5r_package_version
#> [1] ‘1.0.1’
#> 
#> $r5_jar_version
#> [1] "6.9"
#> 
#> $java_version
#> [1] "11.0.18"
#> 
#> $set_memory
#> [1] "-Xmx512m"
#> 
#> $session_info
#> R version 4.2.3 (2023-03-15)
#> Platform: x86_64-pc-linux-gnu (64-bit)
#> Running under: Ubuntu 20.04.6 LTS
#> 
#> Matrix products: default
#> BLAS:   /usr/lib/x86_64-linux-gnu/blas/libblas.so.3.9.0
#> LAPACK: /usr/lib/x86_64-linux-gnu/lapack/liblapack.so.3.9.0
#> 
#> locale:
#>  [1] LC_CTYPE=C.UTF-8          LC_NUMERIC=C             
#>  [3] LC_TIME=C.UTF-8           LC_COLLATE=C             
#>  [5] LC_MONETARY=C.UTF-8       LC_MESSAGES=C.UTF-8      
#>  [7] LC_PAPER=C.UTF-8          LC_NAME=C.UTF-8          
#>  [9] LC_ADDRESS=C.UTF-8        LC_TELEPHONE=C.UTF-8     
#> [11] LC_MEASUREMENT=C.UTF-8    LC_IDENTIFICATION=C.UTF-8
#> 
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base     
#> 
#> other attached packages:
#> [1] r5r_1.0.1
#> 
#> loaded via a namespace (and not attached):
#>  [1] tidyselect_1.2.0   xfun_0.38          purrr_1.0.1        sf_1.0-12         
#>  [5] rJava_1.0-6        testthat_3.1.7     vctrs_0.6.1        generics_0.1.3    
#>  [9] htmltools_0.5.5    yaml_2.3.7         utf8_1.2.3         rlang_1.1.0       
#> [13] e1071_1.7-13       pkgdown_2.0.7      pillar_1.9.0       glue_1.6.2        
#> [17] withr_2.5.0        DBI_1.1.3          lifecycle_1.0.3    ragg_1.2.5        
#> [21] zip_2.2.2          memoise_2.0.1      evaluate_0.20      knitr_1.42        
#> [25] callr_3.7.3        fastmap_1.1.1      ps_1.7.4           curl_5.0.0        
#> [29] class_7.3-21       fansi_1.0.4        Rcpp_1.0.10        KernSmooth_2.23-20
#> [33] backports_1.4.1    classInt_0.4-9     checkmate_2.1.0    cachem_1.0.7      
#> [37] desc_1.4.2         jsonlite_1.8.4     fs_1.6.1           systemfonts_1.0.4 
#> [41] brio_1.1.3         textshaping_0.3.6  digest_0.6.31      processx_3.8.0    
#> [45] dplyr_1.1.1        rprojroot_2.0.3    grid_4.2.3         cli_3.6.1         
#> [49] tools_4.2.3        magrittr_2.0.3     proxy_0.4-27       tibble_3.2.1      
#> [53] whisker_0.4.1      pkgconfig_2.0.3    downlit_0.4.2      data.table_1.14.8 
#> [57] xml2_1.3.3         rmarkdown_2.21     httr_1.4.5         R6_2.5.1          
#> [61] units_0.8-1        compiler_4.2.3    
#>