Generate an r5r situation report to help debug errors

The function reports a list with the following information:

The package version of {r5r} in use.

The installed version of R5.jar.

The Java version in use.

The amount of memory set to Java through the java.parameters option.

The user's Session Info.

Usage

r5r_sitrep()

Value

A list with information of the versions of the r5r package, Java and R5 Jar in use, the memory set to Java and user's Session Info.

Examples

r5r_sitrep()
#> $r5r_package_version
#> [1] ‘2.2.99999’
#> 
#> $r5_jar_version
#> [1] "7.4"
#> 
#> $java_version
#> [1] "21.0.7"
#> 
#> $set_memory
#> [1] "-Xmx2G"
#> 
#> $session_info
#> R version 4.5.1 (2025-06-13)
#> Platform: x86_64-pc-linux-gnu
#> Running under: Ubuntu 22.04.5 LTS
#> 
#> Matrix products: default
#> BLAS:   /usr/lib/x86_64-linux-gnu/openblas-pthread/libblas.so.3 
#> LAPACK: /usr/lib/x86_64-linux-gnu/openblas-pthread/libopenblasp-r0.3.20.so;  LAPACK version 3.10.0
#> 
#> locale:
#>  [1] LC_CTYPE=C.UTF-8          LC_NUMERIC=C             
#>  [3] LC_TIME=C.UTF-8           LC_COLLATE=C             
#>  [5] LC_MONETARY=C.UTF-8       LC_MESSAGES=C.UTF-8      
#>  [7] LC_PAPER=C.UTF-8          LC_NAME=C.UTF-8          
#>  [9] LC_ADDRESS=C.UTF-8        LC_TELEPHONE=C.UTF-8     
#> [11] LC_MEASUREMENT=C.UTF-8    LC_IDENTIFICATION=C.UTF-8
#> 
#> time zone: UTC
#> tzcode source: system (glibc)
#> 
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base     
#> 
#> other attached packages:
#> [1] ggplot2_3.5.2  r5r_2.2.099999
#> 
#> loaded via a namespace (and not attached):
#>  [1] gtable_0.3.6       xfun_0.52          bslib_0.9.0        httr2_1.1.2       
#>  [5] htmlwidgets_1.6.4  processx_3.8.6     rJava_1.0-11       callr_3.7.6       
#>  [9] generics_0.1.4     vctrs_0.6.5        tools_4.5.1        ps_1.9.1          
#> [13] curl_6.4.0         tibble_3.3.0       proxy_0.4-27       fansi_1.0.6       
#> [17] sfheaders_0.4.4    pkgconfig_2.0.3    KernSmooth_2.23-26 data.table_1.17.6 
#> [21] checkmate_2.3.2    RColorBrewer_1.1-3 desc_1.4.3         lifecycle_1.0.4   
#> [25] compiler_4.5.1     farver_2.1.2       brio_1.1.5         textshaping_1.0.1 
#> [29] fontawesome_0.5.3  htmltools_0.5.8.1  class_7.3-23       sass_0.4.10       
#> [33] yaml_2.3.10        pillar_1.10.2      pkgdown_2.1.3      jquerylib_0.1.4   
#> [37] whisker_0.4.1      openssl_2.3.3      classInt_0.4-11    cachem_1.1.0      
#> [41] wk_0.9.4           zip_2.3.3          tidyselect_1.2.1   digest_0.6.37     
#> [45] dplyr_1.1.4        sf_1.0-21          purrr_1.0.4        fastmap_1.2.0     
#> [49] grid_4.5.1         cli_3.6.5          concaveman_1.1.0   magrittr_2.0.3    
#> [53] e1071_1.7-16       withr_3.0.2        scales_1.4.0       backports_1.5.0   
#> [57] rappdirs_0.3.3     rmarkdown_2.29     askpass_1.2.1      ragg_1.4.0        
#> [61] memoise_2.0.1      evaluate_1.0.4     knitr_1.50         testthat_3.2.3    
#> [65] V8_6.0.4           s2_1.1.9           rlang_1.1.6        downlit_0.4.4     
#> [69] Rcpp_1.0.14        glue_1.8.0         DBI_1.2.3          xml2_1.3.8        
#> [73] rstudioapi_0.17.1  jsonlite_2.0.0     R6_2.6.1           rJavaEnv_0.3.0    
#> [77] systemfonts_1.2.3  fs_1.6.6           units_0.8-7       
#>