Initialize Java and Check Version

Sets up Java logging for r5r and ensures Java SE Development Kit 21 is installed.

Usage

start_r5r_java(data_path, temp_dir = FALSE, verbose = FALSE)

Arguments

data_path

A character string. The directory where the log file should be saved.

temp_dir

A logical. Whether the jar file should be saved in a temporary directory. Defaults to FALSE.

verbose

A logical. Whether to show informative messages. Defaults to FALSE.

Value

No return value. The function will stop execution with an error if Java 21 is not found.

Details

This function initializes the Java Virtual Machine (JVM) with a log path for r5r, and checks that the installed Java version is 21. If not, it stops with an informative error message and download links.

See also