Calculate travel time matrix between origin destination pairs

Fast computation of travel time estimates between one or multiple origin destination pairs.

travel_time_matrix(
  r5r_core,
  origins,
  destinations,
  mode = "WALK",
  departure_datetime = Sys.time(),
  max_walk_dist = Inf,
  max_trip_duration = 120L,
  walk_speed = 3.6,
  bike_speed = 12,
  max_rides = 3,
  n_threads = Inf,
  verbose = TRUE
)

Arguments

r5r_core
a rJava object to connect with R5 routing engine

origins, destinations
a spatial sf POINT object, or a data.frame containing the columns 'id', 'lon', 'lat'

mode
string. Transport modes allowed for the trips. Defaults to "WALK". See details for other options.

departure_datetime
POSIXct object. If working with public transport networks, please check calendar.txt within the GTFS file for valid dates.

max_walk_dist
numeric. Maximum walking distance (in meters) for the whole trip. Defaults to no restrictions on walking, as long as max_trip_duration is respected.

max_trip_duration
numeric. Maximum trip duration in minutes. Defaults to 120 minutes (2 hours).

walk_speed
numeric. Average walk speed in km/h. Defaults to 3.6 km/h.

bike_speed
numeric. Average cycling speed in km/h. Defaults to 12 km/h.

max_rides
numeric. The max number of public transport rides allowed in the same trip. Defaults to 3.

n_threads
numeric. The number of threads to use in parallel computing. Defaults to use all available threads (Inf).

verbose
logical. TRUE to show detailed output messages (the default) or FALSE to show only eventual ERROR messages.

Value

A data.table with travel time estimates (in minutes) between origin destination pairs by a given transport mode.

Details

R5 allows for multiple combinations of transport modes. The options include:

Transit modes

TRAM, SUBWAY, RAIL, BUS, FERRY, CABLE_CAR, GONDOLA, FUNICULAR. The option 'TRANSIT' automatically considers all public transport modes available.

Non transit modes

WALK, BICYCLE, CAR, BICYCLE_RENT, CAR_PARK

See also

Examples

# \donttest{
library(r5r)

# build transport network
data_path <- system.file("extdata", package = "r5r")
r5r_obj <- setup_r5(data_path = data_path)

#> Using cached version from C:/Users/user/AppData/Local/Temp/RtmpoxdkBL/temp_libpath7ec26a65e39/r5r/jar/r5r_v4.9.0_20200831.jar

#> 
#> Using cached network.dat from C:/Users/user/AppData/Local/Temp/RtmpoxdkBL/temp_libpath7ec26a65e39/r5r/extdata/network.dat

# load origin/destination points
points <- read.csv(system.file("extdata/poa_hexgrid.csv", package = "r5r"))[1:5,]

mode <- c("WALK", "TRANSIT")
max_walk_dist = Inf
max_trip_duration <- 120L
departure_datetime <- as.POSIXct("13-03-2019 14:00:00",
                                 format = "%d-%m-%Y %H:%M:%S",
                                 tz = "America/Sao_Paulo")

# estimate travel time matrix
df <- travel_time_matrix(r5r_obj,
                         origins = points,
                         destinations = points,
                         mode,
                         departure_datetime,
                         max_walk_dist,
                         max_trip_duration)

stop_r5(r5r_obj)

#> r5r_obj has been successfully stopped.

rJava::.jgc(R.gc = TRUE)

# }