Write a fare structure object to disk

Writes a fare structure object do disk. Fare structure is saved as a collection of .csv files inside a .zip file.

write_fare_structure(fare_structure, file_path)

Arguments

fare_structure

A fare structure object, following the convention set in setup_fare_structure(). This object describes how transit fares should be calculated. Please see the fare structure vignette to understand how this object is structured: vignette("fare_structure", package = "r5r").

file_path

A path to a .zip file. Where the fare structure should be written to.

Value

The path passed to file_path, invisibly.

See also

Examples

library(r5r)

data_path <- system.file("extdata/poa", package = "r5r")
r5r_core <- setup_r5(data_path)
#> Using cached R5 version from /home/runner/work/_temp/Library/r5r/jar/r5-v6.7-all.jar
#> 
#> Using cached network.dat from /home/runner/work/_temp/Library/r5r/extdata/poa/network.dat

fare_structure <- setup_fare_structure(r5r_core, base_fare = 5)

tmpfile <- tempfile("sample_fare_structure", fileext = ".zip")
write_fare_structure(fare_structure, tmpfile)